fix(parse): don't lookup in the cache for non-parsable files
